Font manager created and TrueFont added to project. Manager will return different size fonts based on needs.
Container system implementation began.

Containers will allow for easy positioning in the High Score Screen.
Text Container will allow the Text to be aligned horizontally : LEFT/MIDDLE/RIGHT

Rendering will be combined in the classes for simplicity as the container is strictly a UI element. Current Design is focused on a static representation of items that will not change often.